Added new API: GET: get current orders and add description to swagger Other API: Refactor path imports errorHandler move to Api folder
