+++ web/models/Api/POST/add_job_role.js
@@ -1,6 +1,6 @@
 const ApiError = require("../../../HandleAPI/ApiError");
 const DataApi = require("../../../HandleAPI/DataApi");
-const errorHandler = require('./errorHandler');
+const errorHandler = require('../errorHandler');
 
 const db = require('../../db');
 const add_job_role = async (req, res, next) =>
