+++ web/models/Api/POST/add_storehouse.js
@@ -1,6 +1,6 @@
 const ApiError = require("../../../HandleAPI/ApiError");
 const DataApi = require("../../../HandleAPI/DataApi");
-const errorHandler = require('./errorHandler');
+const errorHandler = require('../errorHandler');
 const db = require('../../db');
 
 const add_storehouse = async (req, res, next) =>
